refactoring homepage and removing old code
