+++ OncoTrials-Frontend/src/pages/Patient/PatientDashboard.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState, useCallback } from 'react'
 import { useNavigate } from 'react-router-dom'
 import supabase from '../../utils/SupabaseClient'
 import { useQuery } from '@tanstack/react-query'
-import TrialCards from './TrialCards'
+import TrialCards from '../../components/shared/TrialCards'
 import SearchTrialsForm from './SearchTrialsForm'
 import HomeNavBar from '../../components/layout/HomeNavBar'
 import PageFooter from '../../components/layout/PageFooter'
